+++ apps/desktop/package.json
@@ -22,6 +22,8 @@
     "@ateam/db": "workspace:*",
     "@ateam/git-core": "workspace:*",
     "@ateam/panes": "workspace:*",
+    "@ateam/protocol": "workspace:*",
+    "@ateam/server": "workspace:*",
     "@electron/notarize": "^3.1.1",
     "@electron/rebuild": "^3.7.1",
     "@types/node": "^22.10.5",
