+++ apps/desktop/tsconfig.node.json
@@ -10,7 +10,6 @@
     "src/main/**/*.ts",
     "src/daemon/**/*.ts",
     "src/preload/**/*.ts",
-    "src/shared/**/*.ts",
     "electron.vite.config.ts"
   ]
 }
