+++ apps/desktop/tsconfig.web.json
@@ -6,5 +6,5 @@
     "jsx": "react-jsx",
     "noEmit": true
   },
-  "include": ["src/renderer/**/*.ts", "src/renderer/**/*.tsx", "src/shared/**/*.ts"]
+  "include": ["src/renderer/**/*.ts", "src/renderer/**/*.tsx"]
 }
