+++ packages/server/src/hooks/board-mcp.ts
