+++ packages/server/src/loops/agent-loop.ts
