+++ packages/server/src/loops/agent-loops.design.md
