+++ packages/server/src/loops/templates.ts
