+++ packages/server/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": { "rootDir": "." },
+  "include": ["src/**/*.ts"]
+}
